save-keys: accept multiple connection
All connections are using the directory_path pointer for opening the file where to write the keys. Without this fix, only the first
connection would have succeeded doing this because the directory_path pointer is freed after the first connection. Now all the
connection can write to the same file.