Avoid O(n^2) copy operations when concatenating netlink responses to one result struct.
When constructing the result, all respones from netlink were concatenated
iterativelly, i.e. for each response, the previously acquired result was
copied to newly allocated place and current response appended to it.
This results in O(n^2) copy operations. Instead, we now check for the
total final length of the result struct and copy the individual
responses to it in one pass, i.e. in O(n) copy operations.
In particular, this issue caused very high CPU usage in memcpy()
function as the result struct is copied over and over. Common way
how to hit the issue is when having 1000+ ipsec routes and 5+
connecting clients a second. In that case, the memcpy() function can
take 50%+ of one CPU thread on a decent CPU and the whole charon daemon
is stuck just reading routes and concatenating them together (connecting
clients are blocked in that particular case as this is done under mutex).