Update SA based on change in priority routing rule
On multihomed hosts, it is common to prioritize one network over the
other(s). For example, an LTE modem over a 2G modem, or a fixed
connection over a satellite connection. Typically, the system has no way
of knowing which interface to prioritize, as the actual network
interfaces looks the same (naming follows the same pattern, etc.). Also,
policies for selecting the preferred interface can be complex and use
multiple inputs (metadata, active checks, configuration options, etc.,
etc.)

Policy routing is commonly used on  multihomed hosts. The routes for the
different interfaces are placed in separate tables, and rules are used
to select which routing table(s) to check for a packet. This PR adds
support for specifying the priority of what I have called a preferred
rule. This rule will always point to the routing table of the preferred
network interface, and the rule is assumed to be updated by the tool
responsible for enforcing the network selection policy.

The feature is currently only available on Linux, and requires MOBIKE to
be enabled and "prefer_best_path" to be set to yes. When a rule with the
given priority is updated (i.e., a NEWRULE is received), a roam event is
triggered. Before the preferred route lookup is performed, a lookup for
the current priority table is done. Routes in this table is placed at
the front of the list that is used to select the preferred route.

The reason for not acting on DELRULE messages, is that I have assumed
that there can only be one preferred rule. Thus, if the rule is deleted,
then there is no preferred network and we can stay with the current one.

Signed-off-by: Kristian Evensen <kristian.evensen@gmail.com>