child-delete: avoid duplicate calls of child_updown
if a CHILD_SA is rekeyed after the beginning of a ike_reauth()
there is a chance that the rekeyed CHILD_SA ends up in state DELETING.
This is due to charon.delete_rekeyed_delay being set to a high
value (like 100) in combination with make_before_break = yes.

In the course of completing ike_reauth() the old IKE_SA
gets destroyed via a call to ike_updown(). This implies calls to
child_updown() but in this situation the fact that the CHILD_SA
is already rekeyed cannot be identified anymore because the state
is already DELETING.