unify child_updown calls when having duplicate QMs
If a Quick mode is initated for a CHILD_SA that is already installed
we can identify this situation and rekey the already installed CHILD_SA.

Otherwise we end up with several CHILD_SAs in state INSTALLED which
means multiple calls of child_updown are done. Unfortunatly
the deduplication code does not call child_updown() (up and down were not
even). Another problem occurs if the newer CHILD_SA has a shorter
lifetime due to rekey fuzz. Then a uneven number of calls to child_updown can be
observed on the responder side.