Allow charon to change group on files before dropping caps
Allow charon to start as a non-root user without CAP_CHOWN and still be
able to change the group on files that need to be accessed by charon
after capabilities have been dropped. This requires user charon starts
as to have access to socket/pidfile directory as well as belong to the
group that charon will run as after dropping capabilities.