libtpmtss: Read public key exponent instead of assuming its value.
Up to now it was assumed that the RSA public key exponent is equal to 2^16+1. Although this is probably true in the most if not all cases, it is not correct according to the TPM specification.

This patch fixes that by reading the exponent from an appropriate of the structure returned by TPM2_ReadPublic.