link-local-ts: Plugin that adds a VIP-based IPv6 link-local address to the remote TS
Windows clients apparently rely on Neighbor Discovery (Router
Advertisements) to install routes for IPv6.  So this alone will not be
enough to make this work.  It requires a tool like `radvd` to actually
send such messages to the clients.

Also, if multiple pools are used, they have to differ in the last 64
bits to avoid assigning duplicate link-local addresses.

 #817