Allow configurable remote traffic selectors
this change allows to customize the previously hard-coded remote traffic selectors. it corresponds to ipsec.conf's `rightsubnet` and charon-cmd's `--remote-ts` parameters, but keeps the comfort of being able to toggle the connection through gnome's gui. 

this does not actually write the newly added "remote-ts" configuration node into the NetworkManager configuration file, but will use an existing value. exposing the config setting in the gui could be done in an amended or followup PR if this is a desired change.

use case: remote firewall appliance wrongly accepts the `0.0.0.0/0` but does not actually route external traffic, leaving the user with a partially working internet connection.